Begin Reorganize API endpoints
