+++ agixt/ApiClient.py
@@ -6,3 +6,28 @@ load_dotenv()
 AGIXT_API_KEY = os.getenv("AGIXT_API_KEY", None)
 DB_CONNECTED = True if os.getenv("DB_CONNECTED", "false").lower() == "true" else False
 ApiClient = AGiXTSDK(base_uri="http://localhost:7437", api_key=AGIXT_API_KEY)
+
+if DB_CONNECTED:
+    from db.Agent import Agent, add_agent, delete_agent, rename_agent, get_agents
+    from db.Chain import Chain
+    from db.Prompts import Prompts
+    from db.History import (
+        get_conversation,
+        delete_history,
+        delete_message,
+        get_conversations,
+        new_conversation,
+        log_interaction,
+    )
+else:
+    from fb.Agent import Agent, add_agent, delete_agent, rename_agent, get_agents
+    from fb.Chain import Chain
+    from fb.Prompts import Prompts
+    from fb.History import (
+        get_conversation,
+        delete_history,
+        delete_message,
+        get_conversations,
+        new_conversation,
+        log_interaction,
+    )
